fixed timer msec overflow.

+++ src/network/Timer.c
@@ -135,7 +135,7 @@ void swTimer_del(swTimer *timer, swTimer_node *tnode)
 
 int swTimer_select(swTimer *timer)
 {
-    int now_msec = swTimer_get_relative_msec();
+    int64_t now_msec = swTimer_get_relative_msec();
     if (now_msec < 0)
     {
         return SW_ERR;
